requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cssselect==1.2.0
DataRecorder==3.6.1
ddddocr==1.5.5
DownloadKit==2.0.4
DrissionPage==4.1.0.8
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==2.1.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pathlib==1.0.1
pillow==10.4.0
protobuf==5.28.2
psutil==6.0.0
pycparser==2.22
PyExecJS==1.5.1
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
requests==2.32.3
requests-file==2.1.0
rfc3986==1.5.0
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sympy==1.13.3
tldextract==5.1.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
